creating exceptions and testing of PersonajeController.

+++ src/main/java/com/prototype/demo/controller/PersonajeController.java
@@ -1,5 +1,7 @@
 package com.prototype.demo.controller;
 
+import com.prototype.demo.excepciones.PersonajeNotFoundException;
+import com.prototype.demo.excepciones.RequestException;
 import com.prototype.demo.model.Personaje;
 import com.prototype.demo.service.IPersonajeService;
 import lombok.AllArgsConstructor;
@@ -9,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.prefs.BackingStoreException;
 
 @RestController
 @RequestMapping("/characters")
@@ -20,23 +23,32 @@ public class PersonajeController {
 
     @GetMapping
     public ResponseEntity<List<Personaje>> obtenerPersonajes(){
-        return ResponseEntity.status(HttpStatus.CREATED).body(iPersonajeService.getPersonajes());
+        return ResponseEntity.ok().body(iPersonajeService.getPersonajes());
     }
 
     @GetMapping(params = "name")
     public ResponseEntity<Personaje> obtenerPersonajePorNombre(@RequestParam("name") String name){
-        return ResponseEntity.status(HttpStatus.CREATED).body(iPersonajeService.getPersonajeByNombre(name));
+        if(!iPersonajeService.existByNombre(name)){
+            throw new PersonajeNotFoundException("EC-003","personaje no existe o parametro de busqueda incorrecto");
+        }
+        return ResponseEntity.status(HttpStatus.FOUND).body(iPersonajeService.getPersonajeByNombre(name));
     }
 
 
     @GetMapping(params = "age")
     public ResponseEntity<List<Personaje>> obtenerPersonajesPorEdad(@RequestParam("age") int age){
-        return ResponseEntity.status(HttpStatus.CREATED).body(iPersonajeService.GetPersonajeByEdad(age));
+        if(!iPersonajeService.existByEdad(age)){
+            throw new PersonajeNotFoundException("EC-003","personaje no existe o parametro de busqueda incorrecto");
+        }
+        return ResponseEntity.status(HttpStatus.FOUND).body(iPersonajeService.GetPersonajeByEdad(age));
     }
 
     @GetMapping(params = "peso")
     public ResponseEntity<List<Personaje>> obtenerPersonajesPorpeso(@RequestParam("peso") float peso){
-        return ResponseEntity.status(HttpStatus.CREATED).body(iPersonajeService.getPersonajeByPeso(peso));
+        if(!iPersonajeService.existByPeso(peso)){
+            throw new PersonajeNotFoundException("EC-003","personaje no existe o parametro de busqueda incorrecto");
+        }
+        return ResponseEntity.status(HttpStatus.FOUND).body(iPersonajeService.getPersonajeByPeso(peso));
     }
 
 
@@ -49,7 +61,11 @@ public class PersonajeController {
 
 
     @PostMapping
-    public ResponseEntity<Personaje> guardarPersonaje (@RequestBody Personaje personaje){
+    public ResponseEntity<Personaje> guardarPersonaje (@RequestBody Personaje personaje) throws RuntimeException{
+        if(personaje.getHistoria()==null || personaje.getEdad()==0 || personaje.getNombre()==null || personaje.getImagen()==null)
+        {
+            throw new RequestException("EC-002","ningún dato puede ser nulo");
+        }
         return ResponseEntity.status(HttpStatus.CREATED).body(iPersonajeService.savePersonaje(personaje));
     }
 
@@ -58,13 +74,17 @@ public class PersonajeController {
 
     @PutMapping
     public ResponseEntity<Personaje> editarPersonaje ( @RequestBody Personaje personaje){
+
         return ResponseEntity.status(HttpStatus.CREATED).body(iPersonajeService.updatePersonaje(personaje));
     }
 
 
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity borrarPersonaje(@PathVariable ("id") Long id){
+    public ResponseEntity<?> borrarPersonaje(@PathVariable ("id") Long id){
+        if(!iPersonajeService.existById(id)){
+            throw new PersonajeNotFoundException("EC-003","personaje no existe o parametro de busqueda incorrecto");
+        }
         iPersonajeService.deletePersonaja(id);
         return ResponseEntity.ok(!iPersonajeService.existById(id));
     }
